data-processing/requirements.txt
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
matplotlib==3.7.2
seaborn==0.12.2
statsbombpy==1.10.0
python-dotenv==1.0.0
jupyter==1.0.0
tqdm==4.66.1 